Main Migrations Class WIP
Created the main migrations class which will manage running migrations.

There was some code that was in both the console application and the Migrations class, so I had the application extend the migrations class and turned the console component class into an interface, which is what it should have been in the first place.